requirements.txt
attrs==24.3.0
gym==0.23.0
h5py==3.13.0
imageio==2.36.1
imageio-ffmpeg==0.5.1
json5==0.12.1
keras==3.8.0
matplotlib==3.9.4
moviepy==1.0.3
numba==0.60.0
numpy==1.26.0
numpy-quaternion==2022.4.3
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
pillow==11.0.0
pygame==2.6.1
pyroomacoustics==0.7.0
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.10.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
tokenizers==0.21.1
toml==0.10.2
tqdm==4.67.1
transformers==4.52.4
yacs==0.1.8
